utils/db_connection.py
import mysql.connector
from config import Config

def get_connection():
    try:
        conn = mysql.connector.connect(
            host=Config.DB_HOST,
            user=Config.DB_USER,
            password=Config.DB_PASSWORD,
            database=Config.DB_NAME,
            port=Config.DB_PORT
        )
        return conn

    except Exception as e:
        print("DB ERROR:", e)
        return None