requirements.txt
streamlit
pandas
mysql-connector-python
requests
python-dotenv
